requirements.txt
docker-py==1.10.6
requests==2.21.0